writing-app-aws-agent/writing-practice/requirements.txt
streamlit==1.30.0
boto3==1.34.11
langchain==0.1.0
pillow==10.1.0
requests==2.31.0 